move TripsLayer and PlotLayer into sample-layers
