+++ demo/src/components/demos/index.js
@@ -2,19 +2,19 @@ import {default as ScatterplotDemo} from './scatterplot';
 import {default as ArcDemo} from './arc';
 import {default as GridDemo} from './grid';
 import {default as ChoroplethDemo} from './choropleth';
-import {default as HeroDemo} from './hero';
+import {default as TripsDemo} from './trips';
 import {default as PlotDemo} from './plot';
 
-class HomeDemo extends HeroDemo {
+class HomeDemo extends TripsDemo {
 
   static get data() {
     return [
       {
-        ...HeroDemo.data[0],
-        url: 'data/hero-data-s.txt'
+        ...TripsDemo.data[0],
+        url: 'data/trips-data-s.txt'
       },
       {
-        ...HeroDemo.data[1],
+        ...TripsDemo.data[1],
         url: 'data/building-data-s.txt'
       }
     ];
@@ -22,7 +22,7 @@ class HomeDemo extends HeroDemo {
 
   static get viewport() {
     return {
-      ...HeroDemo.viewport,
+      ...TripsDemo.viewport,
       longitude: -74.01,
       latitude: 40.707,
       zoom: 14,
@@ -37,7 +37,7 @@ export default {
   ArcDemo,
   GridDemo,
   ChoroplethDemo,
-  HeroDemo,
+  TripsDemo,
   HomeDemo,
   PlotDemo
 };
