+++ demo/src/static/workers/trips-data-decoder.js
