+++ examples/sample-layers/layers/index.js
@@ -1,0 +1,4 @@
+export * from './enhanced-choropleth-layer';
+export * from './plot-layer';
+export * from './trips-layer';
+
