+++ examples/sample-layers/layers/plot-layer/fragment.glsl
