+++ examples/sample-layers/layers/plot-layer/grid-vertex.glsl
