+++ examples/sample-layers/layers/plot-layer/index.js
@@ -1,0 +1,3 @@
+export {default as AxesLayer} from './axes-layer';
+export {default as SurfaceLayer} from './surface-layer';
+export {default as PlotLayer} from './plot-layer';
