+++ examples/sample-layers/layers/plot-layer/label-fragment.glsl
