+++ examples/sample-layers/layers/plot-layer/label-vertex.glsl
