+++ examples/sample-layers/layers/plot-layer/plot-layer.js
