+++ examples/sample-layers/layers/plot-layer/surface-vertex.glsl
