+++ examples/sample-layers/layers/trips-layer/index.js
@@ -1,0 +1,1 @@
+export {default as TripsLayer} from './trips-layer';
