+++ examples/sample-layers/layers/trips-layer/trips-layer-fragment.glsl
