+++ examples/sample-layers/layers/trips-layer/trips-layer-vertex.glsl
