+++ examples/sample-layers/layers/trips-layer/trips-layer.js
@@ -155,3 +155,5 @@ export default class TripsLayer extends Layer {
   }
 
 }
+
+TripsLayer.layerName = 'TripsLayer';
